+++ packages/adp-tooling/src/types.ts
@@ -45,6 +45,8 @@ export interface OnpremApp {
     /** Reference associated with the ID of the base application. */
     reference: string;
     layer?: UI5FlexLayer;
+    fioriId?: string;
+    ach?: string;
     title?: string;
     /** Optional: Application variant change content. */
     content?: Content[];
@@ -101,6 +103,9 @@ export interface ConfigAnswers {
     username: string;
     password: string;
     application: SourceApplication;
+    fioriId?: string;
+    ach?: string;
+    shouldCreateExtProject?: boolean;
 }
 
 export interface AttributesAnswers {